Move helpers.GetTalks to talks.GetTalks
