+++ speaker/speaker.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/devopsdays/devopsdays-cli/helpers"
 	"github.com/devopsdays/devopsdays-cli/model"
+	"github.com/devopsdays/devopsdays-cli/talks"
 	"github.com/fatih/color"
 	survey "gopkg.in/AlecAivazis/survey.v1"
 )
@@ -162,7 +163,7 @@ func CreateSpeaker(speakerName, city, year string) (err error) {
 	if name == true {
 		prompt := &survey.Select{
 			Message: "Choose a talk:",
-			Options: helpers.GetTalks(city, year),
+			Options: talks.GetTalks(city, year),
 		}
 		survey.AskOne(prompt, &talk, nil)
 		color.Yellow("NOT IMPLEMENTED")
