+++ talks/talks.go
@@ -1,0 +1,3 @@
+// Package talks includes the functionality to add, create, edit, remove, and show talks.
+// It also includes supporting and helper functions that are talk-releated.
+package talks
